watchmate/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
PyJWT==2.6.0
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
